SW-4761 Create MyAccountRouter and put views and sub components into scenes directory. Move the observations router to the scenes directory
