+++ src/scenes/MyAccountRouter/AssignNewOwnerModal.tsx
@@ -2,10 +2,10 @@ import { makeStyles } from '@mui/styles';
 import { Theme } from '@mui/material';
 import React from 'react';
 import strings from 'src/strings';
-import Button from '../common/button/Button';
-import Select from '../common/Select/Select';
+import Button from '../../components/common/button/Button';
+import Select from '../../components/common/Select/Select';
 import { OrganizationUser } from 'src/types/User';
-import DialogBox from '../common/DialogBox/DialogBox';
+import DialogBox from '../../components/common/DialogBox/DialogBox';
 
 const useStyles = makeStyles((theme: Theme) => ({
   select: {
