+++ src/scenes/MyAccountRouter/DeleteOrgModal.tsx
@@ -1,26 +1,27 @@
 import React from 'react';
 import strings from 'src/strings';
-import Button from '../common/button/Button';
-import DialogBox from '../common/DialogBox/DialogBox';
+import Button from '../../components/common/button/Button';
+import DialogBox from '../../components/common/DialogBox/DialogBox';
 
-export interface CannotRemoveOrgDialogProps {
+export interface DeleteOrgDialogProps {
   open: boolean;
   onClose: () => void;
   onSubmit: () => void;
+  orgName: string;
 }
 
-export default function CannotRemoveOrgDialog(props: CannotRemoveOrgDialogProps): JSX.Element {
-  const { onClose, open, onSubmit } = props;
+export default function DeleteOrgDialog(props: DeleteOrgDialogProps): JSX.Element {
+  const { onClose, open, onSubmit, orgName } = props;
 
   return (
     <DialogBox
       onClose={onClose}
       open={open}
-      title={strings.CANNOT_REMOVE}
+      title={strings.DELETE_ORGANIZATION}
       size='medium'
       middleButtons={[
         <Button
-          id='cancelCannotRemoveOrg'
+          id='cancelDeleteOrg'
           label={strings.CANCEL}
           priority='secondary'
           type='passive'
@@ -28,14 +29,15 @@ export default function CannotRemoveOrgDialog(props: CannotRemoveOrgDialogProps)
           key='button-1'
         />,
         <Button
-          id='deleteCannotRemoveOrg'
+          id='deleteOrg'
           label={strings.DELETE}
           type='destructive'
+          icon='iconTrashCan'
           onClick={onSubmit}
           key='button-2'
         />,
       ]}
-      message={strings.CANNOT_REMOVE_MSG}
+      message={strings.formatString(strings.DELETE_ORGANIZATION_MSG, orgName)}
       skrim={true}
     />
   );
