+++ src/scenes/MyAccountRouter/LeaveOrganizationModal.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import strings from 'src/strings';
-import Button from '../common/button/Button';
-import DialogBox from '../common/DialogBox/DialogBox';
+import Button from '../../components/common/button/Button';
+import DialogBox from '../../components/common/DialogBox/DialogBox';
 
 export interface LeaveOrganizationDialogProps {
   open: boolean;
