+++ src/scenes/ObservationsRouter/common/AggregatedPlantsStats.tsx
