+++ src/scenes/ObservationsRouter/common/DetailsPage.tsx
