+++ src/scenes/ObservationsRouter/common/SpeciesMortalityRateChart.tsx
