+++ src/scenes/ObservationsRouter/common/SpeciesTotalPlantsChart.tsx
