+++ src/scenes/ObservationsRouter/details/ObservationDetailsRenderer.tsx
