+++ src/scenes/ObservationsRouter/details/index.tsx
