+++ src/scenes/ObservationsRouter/plot/MonitoringPlotPhotos.tsx
