+++ src/scenes/ObservationsRouter/replacePlot/ReplaceObservationPlotModal.tsx
