+++ src/scenes/ObservationsRouter/zone/ObservationPlantingZoneRenderer.tsx
