+++ src/scenes/PeopleRouter/PeopleListView.tsx
@@ -11,8 +11,8 @@ import TfMain from '../../components/common/TfMain';
 import TableCellRenderer from './TableCellRenderer';
 import PageSnackbar from 'src/components/PageSnackbar';
 import RemovePeopleDialog from './RemovePeopleModal';
-import AssignNewOwnerDialog from '../../components/MyAccount/AssignNewOwnerModal';
-import DeleteOrgDialog from '../../components/MyAccount/DeleteOrgModal';
+import AssignNewOwnerDialog from '../MyAccountRouter/AssignNewOwnerModal';
+import DeleteOrgDialog from '../MyAccountRouter/DeleteOrgModal';
 import CannotRemovePeopleDialog from './CannotRemovePeopleModal';
 import { OrganizationService, OrganizationUserService, Response } from 'src/services';
 import { Grid, Theme, useTheme } from '@mui/material';
